requirements.txt
numpy~=1.24.3
pandas~=2.0.1
Pillow~=9.5.0
opencv-contrib-python~=4.7.0.72